switch setPresenterInPodReqMsg to in-file function, not method
