--- bigbluebutton-html5/imports/api/presentation-pods/server/methods.js
@@ -1,6 +1,0 @@
-import { Meteor } from 'meteor/meteor';
-import setPresenterInPodReqMsg from './methods/setPresenterInPodReqMsg';
-
-Meteor.methods({
-  setPresenterInPodReqMsg,
-});
