--- bigbluebutton-html5/imports/api/presentation-pods/server/methods/setPresenterInPodReqMsg.js
@@ -1,25 +1,0 @@
-import { Meteor } from 'meteor/meteor';
-import { check } from 'meteor/check';
-import RedisPubSub from '/imports/startup/server/redis';
-import { extractCredentials } from '/imports/api/common/server/helpers';
-
-export default function setPresenterInPodReqMsg(credentials) { // TODO-- switch to meetingId, etc
-  const REDIS_CONFIG = Meteor.settings.private.redis;
-  const CHANNEL = REDIS_CONFIG.channels.toAkkaApps;
-  const EVENT_NAME = 'SetPresenterInPodReqMsg';
-
-  const { presenterId } = credentials;
-
-  const { meetingId, requesterUserId } = extractCredentials(this.userId);
-
-  check(meetingId, String);
-  check(requesterUserId, String);
-  check(presenterId, String);
-
-  const payload = {
-    podId: 'DEFAULT_PRESENTATION_POD',
-    nextPresenterId: presenterId,
-  };
-
-  RedisPubSub.publishUserMessage(CHANNEL, EVENT_NAME, meetingId, requesterUserId, payload);
-}
